Fixing dummy local claims
