+++ modules/migration/migration-service/src/main/java/org/wso2/carbon/is/migration/service/v530/ClaimManager.java
@@ -113,7 +113,6 @@ public class ClaimManager {
             }
         }
         report.append("\n\n");
-        log.info("end adding local claims");
     }
 
     /**
