+++ modules/migration/migration-service/src/main/java/org/wso2/carbon/is/migration/service/v530/migrator/ClaimDataMigrator.java
@@ -22,19 +22,36 @@ import org.apache.commons.lang.StringUtils;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.wso2.carbon.identity.base.IdentityRuntimeException;
+import org.wso2.carbon.identity.claim.metadata.mgt.dao.CacheBackedLocalClaimDAO;
+import org.wso2.carbon.identity.claim.metadata.mgt.dao.ClaimDialectDAO;
+import org.wso2.carbon.identity.claim.metadata.mgt.dao.LocalClaimDAO;
+import org.wso2.carbon.identity.claim.metadata.mgt.exception.ClaimMetadataException;
+import org.wso2.carbon.identity.claim.metadata.mgt.model.AttributeMapping;
+import org.wso2.carbon.identity.claim.metadata.mgt.model.ClaimDialect;
+import org.wso2.carbon.identity.claim.metadata.mgt.model.LocalClaim;
 import org.wso2.carbon.identity.claim.metadata.mgt.util.ClaimConstants;
 import org.wso2.carbon.identity.core.migrate.MigrationClientException;
 import org.wso2.carbon.identity.core.util.IdentityDatabaseUtil;
 import org.wso2.carbon.identity.core.util.IdentityTenantUtil;
+import org.wso2.carbon.is.migration.internal.ISMigrationServiceDataHolder;
 import org.wso2.carbon.is.migration.service.Migrator;
 import org.wso2.carbon.is.migration.service.v530.ClaimManager;
 import org.wso2.carbon.is.migration.service.v530.SQLConstants;
 import org.wso2.carbon.is.migration.service.v530.bean.Claim;
 import org.wso2.carbon.is.migration.service.v530.bean.Dialect;
 import org.wso2.carbon.is.migration.service.v530.bean.MappedAttribute;
+import org.wso2.carbon.is.migration.service.v540.util.FileBasedClaimBuilder;
+import org.wso2.carbon.is.migration.util.Constant;
 import org.wso2.carbon.is.migration.util.Utility;
+import org.wso2.carbon.user.api.Tenant;
+import org.wso2.carbon.user.api.UserRealm;
+import org.wso2.carbon.user.api.UserStoreException;
+import org.wso2.carbon.user.core.UserCoreConstants;
+import org.wso2.carbon.user.core.claim.inmemory.ClaimConfig;
 
+import javax.xml.stream.XMLStreamException;
 import java.io.FileNotFoundException;
+import java.io.IOException;
 import java.io.PrintWriter;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
@@ -47,6 +64,9 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
+import static org.wso2.carbon.is.migration.util.Constant.MIGRATION_LOG;
+import static org.wso2.carbon.is.migration.util.Constant.SUPER_TENANT_ID;
+
 public class ClaimDataMigrator extends Migrator{
 
     private static Log log = LogFactory.getLog(ClaimDataMigrator.class);
@@ -54,6 +74,14 @@ public class ClaimDataMigrator extends Migrator{
     private boolean isSuccess = true;
     //This is used to record error log
     private int count = 1;
+
+    private static final String CLAIM_CONFIG = "claim-config.xml";
+
+    private ClaimConfig claimConfig;
+
+    private ClaimDialectDAO claimDialectDAO = new ClaimDialectDAO();
+
+    private CacheBackedLocalClaimDAO localClaimDAO = new CacheBackedLocalClaimDAO(new LocalClaimDAO());
 
     @Override
     public void migrate() throws MigrationClientException {
@@ -189,6 +217,8 @@ public class ClaimDataMigrator extends Migrator{
 
                 // Add Local Claims.
                 claimManager.addLocalClaims(claims, report);
+                migrateLocalClaims();
+                log.info("end adding local claims");
 
                 // Add External Claims
                 claimManager.addExternalClaim(claims, report);
@@ -394,5 +424,163 @@ public class ClaimDataMigrator extends Migrator{
             dialectMappedAttributes.put(entry.getKey().replace("@" + tenantDomain, ""), attributes);
             tenantDialectMappedAttributes.put(tenantDomain, dialectMappedAttributes);
         }
+    }
+
+    private void migrateLocalClaims() throws MigrationClientException {
+
+        String filePath = Utility.getDataFilePath(CLAIM_CONFIG, getVersionConfig().getVersion());
+        try {
+            claimConfig = FileBasedClaimBuilder.buildClaimMappingsFromConfigFile(filePath);
+        } catch (IOException | XMLStreamException | UserStoreException e) {
+            String message = "Error while building claims from config file";
+            if (isContinueOnError()) {
+                log.error(message, e);
+            } else {
+                throw new MigrationClientException(message, e);
+            }
+        }
+
+        if (claimConfig.getClaims().isEmpty()) {
+            log.info(MIGRATION_LOG + "No data to migrate related with claim mappings.");
+            return;
+        }
+
+        try {
+            // Migrate super tenant.
+            migrateLocalClaimData(SUPER_TENANT_ID);
+
+            // Migrate other tenants.
+            List<Tenant> tenants = Utility.getTenants();
+            List<Integer> inactiveTenants = Utility.getInactiveTenants();
+            boolean ignoreForInactiveTenants = isIgnoreForInactiveTenants();
+            for (Tenant tenant : tenants) {
+                int tenantId = tenant.getId();
+                if (ignoreForInactiveTenants && inactiveTenants.contains(tenantId)) {
+                    log.info("Skipping claim data migration for inactive tenant: " + tenantId);
+                    continue;
+                }
+                migrateLocalClaimData(tenant.getId());
+            }
+        } catch (UserStoreException | ClaimMetadataException e) {
+            String message = "Error while migrating claim data";
+            if (isContinueOnError()) {
+                log.error(message, e);
+            } else {
+                throw new MigrationClientException(message, e);
+            }
+        }
+    }
+
+    /**
+     * Migrate claim mappings.
+     *
+     * @param tenantId tenant id
+     * @throws UserStoreException     UserStoreException
+     * @throws ClaimMetadataException ClaimMetadataException
+     */
+    private void migrateLocalClaimData(int tenantId) throws UserStoreException, ClaimMetadataException {
+
+        UserRealm realm = ISMigrationServiceDataHolder.getRealmService().getTenantUserRealm(tenantId);
+        String primaryDomainName = realm.getRealmConfiguration().getUserStoreProperty(UserCoreConstants.RealmConfig
+                .PROPERTY_DOMAIN_NAME);
+        if (StringUtils.isBlank(primaryDomainName)) {
+            primaryDomainName = UserCoreConstants.PRIMARY_DEFAULT_DOMAIN_NAME;
+        }
+
+        Set<String> claimDialects = new HashSet<>();
+        for (ClaimDialect claimDialect : claimDialectDAO.getClaimDialects(tenantId)) {
+            claimDialects.add(claimDialect.getClaimDialectURI());
+        }
+
+        Map<String, org.wso2.carbon.user.core.claim.ClaimMapping> externalClaims = new HashMap<>();
+        Set<String> existingLocalClaimURIs = new HashSet<>();
+
+        // Add local claim mappings.
+        for (Map.Entry<String, org.wso2.carbon.user.core.claim.ClaimMapping> entry : claimConfig.getClaims()
+                .entrySet()) {
+
+            String claimURI = entry.getKey();
+            org.wso2.carbon.user.core.claim.ClaimMapping claimMapping = entry.getValue();
+            String claimDialectURI = claimMapping.getClaim().getDialectURI();
+
+            if (ClaimConstants.LOCAL_CLAIM_DIALECT_URI.equals(claimDialectURI)) {
+                if (existingLocalClaimURIs.isEmpty()) {
+                    existingLocalClaimURIs = getExistingLocalClaimURIs(tenantId);
+                }
+                if (existingLocalClaimURIs.contains(claimURI)) {
+                    log.warn(MIGRATION_LOG + "Local claim: " + claimURI + " already exists in the system for" +
+                            " tenant: " + tenantId);
+                    continue;
+                }
+
+                addLocalClaimMapping(tenantId, primaryDomainName, claimURI, claimMapping);
+                existingLocalClaimURIs.add(claimURI);
+            } else {
+                externalClaims.put(entry.getKey(), entry.getValue());
+            }
+        }
+    }
+
+    /**
+     * Get existing local claim URIs.
+     *
+     * @param tenantId tenant id
+     * @return existing claim URI set
+     * @throws ClaimMetadataException ClaimMetadataException
+     */
+    private Set<String> getExistingLocalClaimURIs(int tenantId) throws ClaimMetadataException {
+
+        Set<String> localClaimURIs = new HashSet<>();
+        for (LocalClaim localClaim : localClaimDAO.getLocalClaims(tenantId)) {
+            localClaimURIs.add(localClaim.getClaimURI());
+        }
+        return localClaimURIs;
+    }
+
+    /**
+     * Add local claim mapping.
+     *
+     * @param tenantId          tenant id
+     * @param primaryDomainName primary domain name
+     * @param claimURI          claim URI
+     * @param claimMapping      claim mappings
+     * @throws ClaimMetadataException ClaimMetadataException
+     */
+    private void addLocalClaimMapping(int tenantId, String primaryDomainName, String claimURI, org.wso2.carbon.user
+            .core.claim.ClaimMapping claimMapping) throws ClaimMetadataException {
+
+        List<AttributeMapping> mappedAttributes = new ArrayList<>();
+        if (StringUtils.isNotBlank(claimMapping.getMappedAttribute())) {
+            mappedAttributes.add(new AttributeMapping(primaryDomainName, claimMapping.getMappedAttribute()));
+        }
+        if (claimMapping.getMappedAttributes() != null) {
+            for (Map.Entry<String, String> claimMappingEntry : claimMapping.getMappedAttributes().entrySet()) {
+                mappedAttributes.add(new AttributeMapping(claimMappingEntry.getKey(), claimMappingEntry.getValue()));
+            }
+        }
+
+        Map<String, String> claimProperties = claimConfig.getPropertyHolder().get(claimURI);
+        claimProperties.remove(ClaimConstants.DIALECT_PROPERTY);
+        claimProperties.remove(ClaimConstants.CLAIM_URI_PROPERTY);
+        claimProperties.remove(ClaimConstants.ATTRIBUTE_ID_PROPERTY);
+
+        if (!claimProperties.containsKey(ClaimConstants.DISPLAY_NAME_PROPERTY)) {
+            claimProperties.put(ClaimConstants.DISPLAY_NAME_PROPERTY, "0");
+        }
+        if (claimProperties.containsKey(ClaimConstants.SUPPORTED_BY_DEFAULT_PROPERTY) &&
+                StringUtils.isBlank(claimProperties.get(ClaimConstants.SUPPORTED_BY_DEFAULT_PROPERTY))) {
+            claimProperties.put(ClaimConstants.SUPPORTED_BY_DEFAULT_PROPERTY, "true");
+        }
+        if (claimProperties.containsKey(ClaimConstants.READ_ONLY_PROPERTY) &&
+                StringUtils.isBlank(claimProperties.get(ClaimConstants.READ_ONLY_PROPERTY))) {
+            claimProperties.put(ClaimConstants.READ_ONLY_PROPERTY, "true");
+        }
+        if (claimProperties.containsKey(ClaimConstants.REQUIRED_PROPERTY) &&
+                StringUtils.isBlank(claimProperties.get(ClaimConstants.REQUIRED_PROPERTY))) {
+            claimProperties.put(ClaimConstants.REQUIRED_PROPERTY, "true");
+        }
+
+        LocalClaim localClaim = new LocalClaim(claimURI, mappedAttributes, claimProperties);
+        localClaimDAO.addLocalClaim(localClaim, tenantId);
     }
 }
